Macro r3_portkit::sym::sym_static

pub macro sym_static {
    (
        // A trait item
        $(#[$meta:meta])*
        $vis:vis static $name:ident: SymStatic<$ty:ty>
    ) => { ... },
    (
        // A concrete definition
        $(#[$meta:meta])*
        $vis:vis static $name:ident: SymStatic<$ty:ty> = zeroed!()
    ) => { ... },
}

Define a fn item actually representing a static variable.

Notes regarding accessing the variable in asm!

The variable might actually be placed in an address that is slightly off from what the defined fn item represents. For example, on Thumb targets, the least-significant bit of the fn item’s address is set to indicate that it’s a Thumb function. However, we are actually using it as a variable storage, so the bit must be cleared before use.

The SymStaticExt trait’s methods automatically take care of such situations. When referencing it from assembler code, you must append _ to the symbol name (see the following example).

Examples

#![feature(asm)]
#![feature(naked_functions)]
use r3_portkit::sym::{sym_static, SymStatic, SymStaticExt};
use std::cell::Cell;

struct InteriorMutable(Cell<usize>);
unsafe impl Sync for InteriorMutable {}
unsafe impl r3::utils::ZeroInit for InteriorMutable {}

trait Tr {
    sym_static!(static VAR: SymStatic<InteriorMutable> = zeroed!());
}

impl Tr for u8 {}
impl Tr for u16 {}
impl Tr for &'_ u8 {}
impl<T> Tr for (T,) {}

fn main() {
    // TODO: Replace with the following when const arguments become
    //       inferrable:
    //
    //       let sr1: SymStatic<InteriorMutable, _> = u8::VAR;
    let sr1 = sym_static(u8::VAR);
    sr1.as_ref().0.set(42);
    assert_eq!(sr1.as_ref().0.get(), 42);

    // Each instantiation gets a unique storage.
    let sr2 = sym_static(u16::VAR);
    assert_eq!(sr2.as_ref().0.get(), 0);
    sr2.as_ref().0.set(84);
    assert_eq!(sr1.as_ref().0.get(), 42);
    assert_eq!(sr2.as_ref().0.get(), 84);

    // ...however, types only differing by lifetime parameters do not
    // get a unique storage.
    fn inner<'a>(_: &'a mut u8) {
        let sr1 = sym_static(<&'static u8>::VAR);
        let sr2 = sym_static(<&'a u8>::VAR);
        sr1.as_ref().0.set(1);
        sr2.as_ref().0.set(2);
        assert_eq!(sr1.as_ref().0.get(), 2);
        assert_eq!(sr2.as_ref().0.get(), 2);
    }
    inner(&mut 42);

    // Since it appears as a `fn` item, it can be passed to `asm!` by a
    // `sym` input.
    let got_value: usize;
    #[cfg(target_arch = "x86_64")]
    #[cfg(target_os = "linux")]
    unsafe {
        asm!("
            mov {0}, qword ptr [rip + {1}_@GOTPCREL]
            mov {0}, qword ptr [{0}]",
            out(reg) got_value, sym u8::VAR,
        );
    }
    #[cfg(target_arch = "x86_64")]
    #[cfg(not(target_os = "linux"))]
    unsafe { asm!("mov {}, qword ptr [rip + {}_]", out(reg) got_value, sym u8::VAR) };
    #[cfg(target_arch = "aarch64")]
    #[cfg(target_os = "macos")]
    unsafe {
        asm!("
            adrp {0}, {1}_@PAGE
            ldr {0}, [{0}, {1}_@PAGEOFF]
            ",
            out(reg) got_value, sym u8::VAR
        );
    }
    #[cfg(target_arch = "aarch64")]
    #[cfg(not(target_os = "macos"))]
    unsafe {
        asm!("
            adrp {0}, :got:{1}_
            ldr {0}, [{0}, #:got_lo12:{1}_]
            ldr {0}, [{0}]
            ",
            out(reg) got_value, sym u8::VAR
        )
    };
    assert_eq!(got_value, 42);
}