Type Definition r3_portkit::sym::SymStatic

type SymStatic<T, const ALIGN: usize> = unsafe extern "C" fn(_: &'static __UnsafeSymStaticMarker<T, ALIGN>) -> !;

Represents a value of &'static T. The functions of this type are defined by sym_static!.

The ALIGN parameter and its existence or non-existence are exempt from the API stability guarantee. For now, use sym_static() to coerce a value into this type without explicitly specifying its parameters.