Module r3_portkit::tickful

Implements the core algorithm for tickful timing.

Structs

TickfulCfg
The precomputed parameters for the tickful implementation of r3::kernel::PortTimer.

TickfulOptions
The parameters of the tickful timing algorithm.

TickfulStateCore
The internal state of the tickful implementation of r3::kernel::PortTimer.

Enums

CfgError
Error type for TickfulCfg::new.

Traits

TickfulStateTrait
Operations implemented by all valid instantiations of TickfulState.

Type Definitions

TickfulState
Instantiates the optimal version of TickfulStateCore using a given TickfulCfg. All instances implement TickfulStateTrait.